pyproject.toml
[tool.poetry]
name = "nft_meetup"
version = "0.1.0"
description = ""
authors = ["pythoninthegrass <4097471+pythoninthegrass@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.0.1"
ipython = "^8.0.1"

[tool.poetry.dev-dependencies]
icecream = "^2.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
